app/generator/code/requirements.txt
psycopg2-binary
protobuf==5.29.1
